+++ app/src/main/AndroidManifest.xml
@@ -18,12 +18,19 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".admin.mahasiswa.ReadMahasiswaActivity" />
-        <activity android:name=".admin.dosen.DosenActivity" />
-        <activity android:name=".admin.mahasiswa.AddMahasiswaActivity" />
-        <activity android:name=".admin.mahasiswa.DetailMahasiswaActivity" />
-        <activity android:name=".admin.dosen.DetailDosen" />
-        <activity android:name=".admin.dosen.AddDosen"></activity>
+        <activity android:name=".main.admin.mahasiswa.ReadMahasiswaActivity" />
+        <activity android:name=".main.admin.dosen.DosenActivity" />
+        <activity android:name=".main.admin.mahasiswa.AddMahasiswaActivity" />
+        <activity android:name=".main.admin.mahasiswa.DetailMahasiswaActivity" />
+        <activity android:name=".main.admin.dosen.DetailDosen" />
+        <activity android:name=".main.admin.dosen.AddDosen" />
+        <activity android:name=".main.user.MahasiswaUserActivity" />
+        <activity android:name=".main.user.DosenUserActivity" />
+        <activity android:name=".main.user.MatakuliahUserActivity" />
+        <activity android:name=".main.user.HomeUserActivity" />
+        <activity android:name=".main.admin.matakuliah.MatakuliahActivity" />
+        <activity android:name=".main.admin.matakuliah.AddMatakuliahActivity" />
+        <activity android:name=".main.admin.matakuliah.DetailMatakuliah"></activity>
     </application>
 
 </manifest>
