+++ app/src/main/java/com/ghozay19/praditaapps/main/admin/adapter/DosenAdapter.java
@@ -1,4 +1,4 @@
-package com.ghozay19.praditaapps.admin.adapter;
+package com.ghozay19.praditaapps.main.admin.adapter;
 
 import android.content.Context;
 import android.content.Intent;
@@ -9,8 +9,8 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.ghozay19.praditaapps.R;
-import com.ghozay19.praditaapps.admin.dosen.DetailDosen;
-import com.ghozay19.praditaapps.admin.mahasiswa.DetailMahasiswaActivity;
+import com.ghozay19.praditaapps.main.admin.dosen.DetailDosen;
+import com.ghozay19.praditaapps.main.admin.mahasiswa.DetailMahasiswaActivity;
 import com.ghozay19.praditaapps.model.Dosen;
 import com.ghozay19.praditaapps.model.Mahasiswa;
