+++ app/src/main/java/com/ghozay19/praditaapps/main/user/MahasiswaUserActivity.java
@@ -1,0 +1,15 @@
+package com.ghozay19.praditaapps.main.user;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.ghozay19.praditaapps.R;
+
+public class MahasiswaUserActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_mahasiswa_user);
+    }
+}
