+++ app/src/main/res/layout/activity_dosen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".admin.mahasiswa.ReadMahasiswaActivity">
+    tools:context=".main.admin.mahasiswa.ReadMahasiswaActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
